Refactor data pipeline and update imports for new dataset structure
- Move dataset classes from input_pipeline.py to data_pipeline.py
- Remove input_pipeline.py and update all imports to use data_pipeline
- Update task and inference modules to use new dataset class locations
- Simplify CFCDataset by removing unused arguments
